Manual and delete reports as admin

+++ html/report_overview.php
@@ -11,27 +11,26 @@ $variables = array (
     'secured' => true,
 );
 
-$user = $_SESSION ['guardian_userid'];
-$usersEngine = get_engine(get_engine_of_user($user));
-
-if (isset ( $_POST ['delete'] )) {
-    $delete_report_uuid = trim ( $_POST ['delete'] );
-    if(delete_report ( $delete_report_uuid )){
-        $variables ['successMessage'] = "Bericht gelöscht";
+if(isset($_SESSION ['guardian_userid'])){
+    $user = $_SESSION ['guardian_userid'];
+    $usersEngine = get_engine(get_engine_of_user($user));
+    
+    if (isset ( $_POST ['delete'] )) {
+        $delete_report_uuid = trim ( $_POST ['delete'] );
+        if(delete_report ( $delete_report_uuid )){
+            $variables ['successMessage'] = "Bericht gelöscht";
+        } else {
+            $variables ['alertMessage'] = "Bericht konnte nicht gelöscht werden";
+        }
+    }
+    
+    if($usersEngine->name == 'Geschäftszimmer'){
+        $variables ['reports'] = get_reports();
     } else {
-        $variables ['alertMessage'] = "Bericht konnte nicht gelöscht werden";
+        $variables ['reports'] = get_filtered_reports($usersEngine->uuid);
+        $variables ['infoMessage'] = "Es werden nur Wachberichte angezeigt, die Ihrem Zug zugewiesen wurden";
     }
 }
-
-if($usersEngine->name == 'Geschäftszimmer'){
-    $variables ['reports'] = get_reports();
-} else {
-    $variables ['reports'] = get_filtered_reports($usersEngine->uuid);
-    $variables ['infoMessage'] = "Es werden nur Wachberichte angezeigt, die Ihrem Zug zugewiesen wurden";
-}
-
-
-
 
 renderLayoutWithContentFile ( "reportOverview_template.php", $variables );
